Moved Azure Entra classes to io.kafbat.ui.config.auth.azure
